Attempted to implement the 2D Kalman filter in Haskell
However, I don't have a good enough grasp on things to understand how
types can or can't be transformed into one another. e.g. how can a
constant matrix of type noiseCov which is a specific instance of a M44
be 'cast' into an M44 result if its used in a function.